adding in dataloader.  still needs tests
